feat: доработать контроллеры по анкете; добавить модуль по созданию изображений и контроллер по любимцам (еще сырые)

+++ prerevolutionary-tinder-server/src/main/java/ru/liga/prerevolutionarytinderserver/mapper/ProfileMapper.java
@@ -8,12 +8,15 @@ import ru.liga.prerevolutionarytinderserver.model.Profile;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+/**
+ * Класс-маппер объекта анкеты к таблице profile
+ */
 @Component
 public class ProfileMapper implements RowMapper<Profile> {
     @Override
     public Profile mapRow(ResultSet rs, int rowNum) throws SQLException {
         return new Profile(
-                rs.getInt("user_id"),
+                rs.getLong("user_id"),
                 rs.getString("name"),
                 GenderEnum.valueOf(rs.getString("gender")),
                 rs.getString("description"),
